+++ gpytorch/mlls/__init__.py
@@ -8,6 +8,7 @@ from .deep_predictive_log_likelihood import DeepPredictiveLogLikelihood
 from .exact_marginal_log_likelihood import ExactMarginalLogLikelihood
 from .gamma_robust_variational_elbo import GammaRobustVariationalELBO
 from .inducing_point_kernel_added_loss_term import InducingPointKernelAddedLossTerm
+from .kl_gaussian_added_loss_term import KLGaussianAddedLossTerm
 from .leave_one_out_pseudo_likelihood import LeaveOneOutPseudoLikelihood
 from .marginal_log_likelihood import MarginalLogLikelihood
 from .noise_model_added_loss_term import NoiseModelAddedLossTerm
@@ -40,6 +41,7 @@ __all__ = [
     "ExactMarginalLogLikelihood",
     "InducingPointKernelAddedLossTerm",
     "LeaveOneOutPseudoLikelihood",
+    "KLGaussianAddedLossTerm",
     "MarginalLogLikelihood",
     "NoiseModelAddedLossTerm",
     "PredictiveLogLikelihood",
